moved some constatns to helpers
